requirements.txt
git+https://github.com/wtsi-hgi/python-common.git@5e137b726da1c6259a32cd116981209d62ccb3cb#egg=hgicommon
bidict
docker-py
stopit